fastapi/requirements.txt
aiofile==3.7.1
aiofiles==0.7.0
aiohttp==3.7.4.post0
aiolimiter==1.0.0b1
aioredis==2.0.0
aiosmtplib==1.1.6
aiosqlite==0.17.0
anyio==3.3.0
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
background==0.2.1
bcrypt==3.2.0
blinker==1.4
caio==0.9.3
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
cycler==0.10.0
data==0.4
databases==0.4.3
decorator==5.0.9
Deprecated==1.2.12
dnspython==2.1.0
elasticsearch==7.14.0
email-validator==1.1.3
fakeredis==1.6.0
fastapi==0.68.0
fastapi-mail==0.4.1
fastapi-users==6.1.2
fastapi-utils==0.2.1
funcsigs==1.0.2
furl==2.1.2
greenlet==1.1.1
h11==0.12.0
httpcore==0.13.6
httpx==0.19.0
idna==3.2
incoming==0.3.1
iniconfig==1.1.1
Jinja2==3.0.1
jmespath==0.10.0
kiwisolver==1.3.1
logging-tree==1.9
loguru==0.5.3
logx==0.1.1579232358
makefun==1.11.3
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
migra==3.0.1621480950
multidict==5.1.0
Naked==0.1.31
numpy==1.21.2
orderedmultidict==1.0.1
packaging==21.0
pandas==1.3.2
passlib==1.7.4
pendulum==2.1.2
pexpect==4.8.0
Pillow==8.3.1
piplapis-python==5.2.2
pluggy==0.13.1
psycopg2-binary==2.9.1
ptyprocess==0.7.0
py==1.10.0
pycparser==2.20
pydantic==1.8.2
PyJWT==2.1.0
pyparsing==2.4.7
pytest==6.2.4
pytest-asyncio==0.15.1
python-dateutil==2.8.2
python-dotenv==0.19.0
python-multipart==0.0.5
pytz==2021.1
pytzdata==2020.1
PyYAML==5.4.1
redis==3.5.3
requests==2.26.0
response==0.5.0
results==0.1.1588569394
rfc3986==1.5.0
schemainspect==3.0.1616029793
scipy==1.7.1
selenium==3.141.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.3.24
sqlbag==0.1.1617247075
sse-starlette==0.7.2
starlette==0.14.2
status==0.2.5
Strip==0.0.1.dev14
subject==0.1.0
tabulate==0.8.9
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.6
username==2.0.2
uvicorn==0.14.0
win32-setctime==1.0.3
wrapt==1.12.1
yarl==1.6.3
